Work in progress:
The edit-menu (copy) is not shown anymore. This should show on a normal Tap.
Text is selectable in bubbles. If this is disables, URLs are not clickable anymore.
Numbers are called through NumberBay. contactsId should be determined + there is a bug where the inputToolbar is shown over the callView (see comments).

The contactsId should be a property of the conversationViewController so it isn't determined each time. (The contact-name should display on the top of the screen)
The PhoneNumber-class should be used.